+++ R/TaskDens.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' This task specializes [TaskUnsupervised] for density estimation problems.
-#' The target column is assumed to be numeric.
+#' The data in `backend` should be a numeric vector or a one column matrix-like object.
 #' The `task_type` is set to `"density"`.
 #'
 #' Predefined tasks are stored in the [dictionary][mlr3misc::Dictionary] [mlr_tasks].
@@ -14,37 +14,28 @@
 #' @family Task
 #' @export
 #' @examples
-#' task = TaskDens$new("precip", backend = data.frame(target = precip), target = "target")
+#' task = TaskDens$new("precip", backend = precip)
 #' task$task_type
-#' task$truth()
 TaskDens = R6::R6Class("TaskDens",
   inherit = TaskUnsupervised,
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
-    #'
-    #' @param target (`character(1)`)\cr
-    #'   Name of the target column.
-    initialize = function(id, backend, target) {
-      super$initialize(id = id, task_type = "dens", backend = backend)
-      assert_subset(target, self$col_roles$feature)
-      self$col_roles$target = target
-      self$col_roles$feature = setdiff(self$col_roles$feature, target)
+    #' @param backend ([DataBackend])\cr
+    #' Either a [DataBackend], a matrix-like object, or a numeric vector.
+    #' If weights are used then two columns expected, otherwise one column. The weight column
+    #' must be clearly specified (via `[Task]$col_roles`) or the learners will break.
+    initialize = function(id, backend) {
 
-
-      type = subset(self$col_info, id == target, "type")
-      if (!(type %in% c("integer", "numeric"))) {
-        stop("Target column '%s' must be numeric", target)
+      if (test_numeric(backend)) {
+        backend = data.frame(x = backend)
+      } else if (test_class(backend, "DataBackend")) {
+        assert_numeric(backend$ncol, lower = 2, upper = 3)
+      } else {
+        assert_numeric(ncol(backend), lower = 1, upper = 2)
       }
-    },
 
-    #' @description
-    #' Returns the target column for specified `row_ids`, this is unsupervised so should not be
-    #' thought of as a 'true' prediction.
-    #' Defaults to all rows with role "use".
-    #' @return `numeric()`.
-    truth = function(rows = NULL) {
-      self$data(rows, cols = self$target_names)[[1]]
+      super$initialize(id = id, task_type = "dens", backend = backend)
     }
   )
 )