+++ tests/testthat/test_mlr_task_generators_simdens.R
@@ -1,5 +1,3 @@
-context("mlr_task_generators_simdens")
-
 test_that("mlr_task_generators_simdens", {
   gen = mlr_task_generators$get("simdens")
   task = gen$generate(10)